Changed to edit stock

+++ backend/analytics-app/src/main/java/com/example/analyticsapp/portfolio/StockRequestDTO.java
@@ -21,7 +21,7 @@ public class StockRequestDTO {
     public double getProportion() {
         return proportion;
     }
-    public void setProportion(float proportion) {
+    public void setProportion(double proportion) {
         this.proportion = proportion;
     }
